Handle cakephp 5.3 deprecations
