+++ plugins/json-ld-view/src/JsonLdViewPlugin.php
@@ -1,19 +1,20 @@
 <?php
 declare(strict_types=1);
 
-namespace MixerApi\HalView;
+namespace MixerApi\JsonLdView;
 
 use Cake\Core\BasePlugin;
+use Cake\Core\Configure;
 use Cake\Core\PluginApplicationInterface;
 
-class Plugin extends BasePlugin
+class JsonLdViewPlugin extends BasePlugin
 {
     /**
      * Plugin name.
      *
-     * @var string|null
+     * @var string
      */
-    protected ?string $name = 'MixerApi/HalView';
+    protected ?string $name = 'MixerApi/JsonLdView';
 
     /**
      * Console middleware
@@ -50,5 +51,21 @@ class Plugin extends BasePlugin
     public function bootstrap(PluginApplicationInterface $app): void
     {
         parent::bootstrap($app);
+
+        try {
+            Configure::load('jsonld_config');
+        } catch (\Exception $e) {
+        }
+
+        if (Configure::read('JsonLdView') === null) {
+            Configure::write('JsonLdView', [
+                'isHydra' => false,
+                'schemaUrl' => 'http://schema.org',
+                'vocabUrl' => '/vocab',
+                'title' => 'API Documentation',
+                'description' => '',
+                'entrypointUrl' => '/',
+            ]);
+        }
     }
 }
