+++ plugins/jwt-auth/src/JwtAuthPlugin.php
@@ -1,0 +1,54 @@
+<?php
+declare(strict_types=1);
+
+namespace MixerApi\JwtAuth;
+
+use Cake\Core\BasePlugin;
+
+/**
+ * Plugin for JwtAuth
+ */
+class JwtAuthPlugin extends BasePlugin
+{
+    /**
+     * Plugin name.
+     *
+     * @var string
+     */
+    protected ?string $name = 'MixerApi/JwtAuth';
+
+    /**
+     * Do bootstrapping or not
+     *
+     * @var bool
+     */
+    protected bool $bootstrapEnabled = false;
+
+    /**
+     * Console middleware
+     *
+     * @var bool
+     */
+    protected bool $consoleEnabled = false;
+
+    /**
+     * Enable middleware
+     *
+     * @var bool
+     */
+    protected bool $middlewareEnabled = false;
+
+    /**
+     * Register container services
+     *
+     * @var bool
+     */
+    protected bool $servicesEnabled = false;
+
+    /**
+     * Load routes or not
+     *
+     * @var bool
+     */
+    protected bool $routesEnabled = false;
+}
